Use own GH_TOKEN secret, not reserved GITHUB_TOKEN
GITHUB_TOKEN is a reserved environment variable in GitHub Actions,
presumably for automatically populating it with an access token defined
within the repository settings, so trying to define it in repository
Settings > Secrets fails. Here we want to use a personal access token
from a GitHub user account, so we're defining our own GH_TOKEN variable
in Settings > Secrets and then assigning it to the GITHUB_TOKEN variable
that AutoPub expects.